Fixed migration_functions

+++ django_db_views/migration_functions.py
@@ -7,21 +7,26 @@ class ViewMigration(object):
         self.view_definition = view_definition
         self.table_name = table_name
 
+
+@deconstructible
+class ForwardViewMigration(ViewMigration):
+
     def __call__(self, apps, schema_editor):
-        schema_editor.execute(
-            'DROP VIEW IF EXISTS %s;' % self.table_name
-        )
         if self.view_definition:
             schema_editor.execute(
-                """CREATE VIEW %s AS %s;""" % (self.table_name, self.view_definition)
+                """CREATE OR REPLACE VIEW %s AS %s;""" % (self.table_name, self.view_definition)
             )
 
 
 @deconstructible
-class ForwardViewMigration(ViewMigration):
-    pass
-
-
-@deconstructible
 class BackwardViewMigration(ViewMigration):
-    pass
+
+    def __call__(self, apps, schema_editor):
+        if self.view_definition:
+            schema_editor.execute(
+                """CREATE OR REPLACE VIEW %s AS %s;""" % (self.table_name, self.view_definition)
+            )
+        else:
+            schema_editor.execute(
+                'DROP VIEW IF EXISTS %s;' % self.table_name
+            )
